♻️ refactor: 시스템 환경 확인 로직 위임
